Add POST /commutes
Adds POST endpoint to parse a JSON body as a Commute object. Does not
save the commute to the database yet.

I don't think this code is as good as it could be, but it's better than
I expected to have come up with in the short amount of time I've looked
at it. I have a feeling I don't need to implement ToJson for Commute and
can instead implement Encodable on Commute and Timestamp, but I wasn't
able to get it working. I think I'll revisit it.

The rust_serialize docs [mention an alternative][1] library, [Serde][2],
which I want to check out as well.

[1]: http://doc.rust-lang.org/rustc-serialize/rustc_serialize/json/index.html#the-status-of-this-library
[2]: https://crates.io/crates/serde